requirements.txt
absl-py==2.1.0
anaconda-anon-usage @ file:///C:/b/abs_c3w_h1zzjg/croot/anaconda-anon-usage_1710965204622/work
archspec @ file:///croot/archspec_1709217642129/work
asttokens==2.4.1
astunparse==1.6.3
blinker==1.9.0
boltons @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/boltons_1699480450092/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
captum==0.7.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///C:/b/abs_90yq4lmu83/croot/cffi_1726856448345/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama==0.4.6
comm==0.2.1
conda @ file:///D:/bld/conda_1732552622811/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///C:/b/abs_831_pypuuz/croot/conda-package-handling_1718138287790/work
conda_package_streaming @ file:///C:/b/abs_fe6zb0_j8_/croot/conda-package-streaming_1718136098642/work
contourpy==1.3.1
cryptography @ file:///C:/b/abs_35g500qir4/croot/cryptography_1724940575116/work
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
executing==2.0.1
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.3
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
libmambapy @ file:///C:/b/abs_aam1usey89/croot/mamba-split_1725563515831/work/libmambapy
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
menuinst @ file:///C:/b/abs_76jwd47ck1/croot/menuinst_1723567618942/work
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
onnx==1.17.0
onnx2pytorch==0.5.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==23.2
pandas==2.2.3
parso==0.8.3
pillow==11.0.0
platformdirs==4.2.0
pluggy @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pluggy_1699472504117/work
prompt-toolkit==3.0.43
protobuf==5.29.1
psutil==5.9.8
pure-eval==0.2.2
pycosat @ file:///C:/b/abs_5csdern___/croot/pycosat_1714513102923/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.17.2
pyparsing==3.2.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.8.2
pytz==2024.2
pywin32==306
pyzmq==25.1.2
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
rich==13.9.4
ruamel.yaml @ file:///C:/b/abs_0cunwx_ww6/croot/ruamel.yaml_1727980181547/work
ruamel.yaml.clib @ file:///C:/b/abs_5fk8zi6n09/croot/ruamel.yaml.clib_1727769837359/work
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm @ file:///C:/b/abs_77wju137gk/croot/tqdm_1724853945787/work
traitlets==5.14.1
truststore @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/truststore_1701881385424/work
typing_extensions==4.12.2
tzdata==2024.2
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wrapt==1.17.0
zstandard @ file:///C:/b/abs_82dff2ths3/croot/zstandard_1728569207191/work
